r/tests: extracted a common fixture to build stm tests
